requirements.txt
praw>=7.8.1
vaderSentiment==3.3.2
python-dotenv==1.0.0
requests==2.32.3

